Fix key distribution bug in compatible mode.
Because of accumulated rounding error some keys were mapped
to the different server than with Cache::Memcached.  Now
we round toward zero, and depend on the fact that the error
can't be too big to cover the whole interval.

Also add script/c-m-compat.pl to compare key distribution
with Cache::Memcached.