Fix t/command.t on some Solaris systems where it fails.
The problem was that I used an old name MAX_IOVEC for the max number
of struct iovec passed to writev()/sendmsg().  When it wasn't defined
a value of 1024 was used.  This was too much for some Solaris
distributions, which returned EINVAL in this case (other systems either
support this number natively, or compensate for it in libc).

Now we use either sysconf(_SC_IOV_MAX), or IOV_MAX, or safe minimum of
16 (64 on Win32).

Thanks to joyent.com for providing necessary Solaris traces.