Fix magic handling (RT#57150, RT#67106).
magic.t was failing because to avoid copying we pass around pointers
to PV internal storage.  In the case of magic value such storage is
created on the first call to mg_get() and is valid until the next call
to mg_get() on the same SV.  $memd->set($scalar2, $scalar2) passes the
same SV as the first and the second argument, so mg_get() was called
on it twice rendering pointer to the key storage invalid.

Thanks to listposter at gmail.com who finally provided the instructions
on how to reproduce the problem (it never failed on Fedora for me, still
perhaps I could try harder from the start, but users should do their
part too ;)).